refactor(LottoNumberGenerationService,LottoPurchaseService,LottoResultCalculationService,LottoWinningNumberService) : LottoService를 4개의 서비스로 분리 리팩토링
