+++ src/main/java/lotto/Application.java
@@ -6,15 +6,21 @@ import lotto.adapter.RandomLottoNumbersProvider;
 import lotto.controller.LottoGameController;
 import lotto.port.InputPort;
 import lotto.port.OutputPort;
-import lotto.service.LottoService;
+import lotto.service.LottoNumberGenerationService;
+import lotto.service.LottoPurchaseService;
+import lotto.service.LottoResultCalculationService;
+import lotto.service.LottoWinningNumberService;
+
 
 public class Application {
     public static void main(String[] args) {
         InputPort inputPort = new ConsoleInputAdapter();
         OutputPort outputPort = new ConsoleOutputAdapter();
-        LottoService lottoService = new LottoService(new RandomLottoNumbersProvider());
-
-        LottoGameController lottoGameController = new LottoGameController(inputPort,outputPort, lottoService);
+        LottoNumberGenerationService lottoNumberGenerationService = new LottoNumberGenerationService(new RandomLottoNumbersProvider());
+        LottoPurchaseService lottoPurchaseService = new LottoPurchaseService();
+        LottoResultCalculationService lottoResultCalculationService = new LottoResultCalculationService();
+        LottoWinningNumberService lottoWinningNumberService = new LottoWinningNumberService();
+        LottoGameController lottoGameController = new LottoGameController(inputPort,outputPort,lottoNumberGenerationService,lottoPurchaseService,lottoResultCalculationService,lottoWinningNumberService);
 
         lottoGameController.playGame();
     }
